+++ src/main.js
@@ -1,6 +1,8 @@
 // Import farmOS map instance factory function.
 import createInstance from './instance/instance';
 
+import namedBehaviors from './behavior';
+
 // Import farmOS-map CSS.
 import './styles.css';
 
@@ -15,11 +17,17 @@ window.farmOS.map = {
   // Initialize an array of farmOS map instances.
   instances: [],
 
-  // Initialize map behaviors.
+  // Behaviors which can be added on-demand by name via `instance.addBehavior`
   // These are objects with an attach() method that will be called when a map is
   // created. They receive the instance object so that they can perform
   // modifications to the map using instance methods.
-  behaviors: {},
+  namedBehaviors,
+
+  // Map behaviors which will be automatically attached to all created map
+  // instances.
+  behaviors: {
+    rememberLayer: namedBehaviors.rememberLayer,
+  },
 
   // Create a new farmOS map attached to a target element ID and add it to the
   // global instances array.
@@ -29,20 +37,10 @@ window.farmOS.map = {
     // Add the instance to the array.
     this.instances.push(instance);
 
-    // Attach behaviors from farmOS.map.behaviors to the instance.
-    // Sort by an optional weight value on each behavior.
-    const behaviors = Object.keys(this.behaviors).map(i => this.behaviors[i]);
-    behaviors.sort((first, second) => {
-      const firstWeight = first.weight || 0;
-      const secondWeight = second.weight || 0;
-      if (firstWeight === secondWeight) {
-        return 0;
-      }
-      return (firstWeight < secondWeight) ? -1 : 1;
-    });
-    behaviors.forEach((behavior) => {
-      instance.attachBehavior(behavior);
-    });
+    // Attach behaviors from farmOS.map.behaviors to the instance in order of weight - if any.
+    instance.defaultBehaviorsAttached = instance.attachBehaviorsByWeight(
+      Object.values(this.behaviors),
+    );
 
     // Return the instance.
     return instance;