+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Check in diff of `package.json`/`package-lock.json` from moving to NPM v7
 - Update all dependencies.
 - Update eslint to allow import() within code.
+- Remove redundant `defaults.behaviors` in favor of existing `farmOS.map.behaviors` mechanism for specifying the default behaviors.
+- Make instance behavior methods `async`.
+- Lazily load behaviors as separate Webpack chunks.
 
 ### Fixed